service: implement AlwaysConnectedTechnologies option
Add a new function always_connect which returns true if a service type
is in the list of technologies to always autoconnect. Call it at various
points in the auto_connect_service function to ensure the service type
always connects regardless of any preffered technologies settings and
correctly handles session logic.